Move train.py args to own file
This halves the size of the file, making it a lot more readable.
